# openapi/ontology.yaml
openapi: 3.1.0
info:
  title: Ontology OpenAPI
  description: Ontology OpenAPI Specification
  version: 0.0.1
servers:
  - url: http://localhost:4242
    description: Local development server
...
paths:
  "/v1/ontologies":
    get:
      tags:
        - Ontology
      summary: List Ontologies
      operationId: listOntologies
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
      responses:
        "200":
          description: Visible Ontologies
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/OntologyPage"
    post:
      tags:
        - Ontology
      summary: Create an Ontology
      operationId: createOntology
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateOntologyReq"
      responses:
        "201":
          description: Created Ontology
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Ontology"
...
  /v1/ontologies/{ontologyId}/object-types/{objectTypeId}/object-instances:
    get:
      tags:
        - ObjectInstance
      summary: List Object Instances
      operationId: listObjectInstances
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Object Instance Page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectInstancePage'
    post:
      tags:
        - ObjectInstance
      summary: Create an Object Instance
      operationId: createObjectInstance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateObjectInstanceReq'
      responses:
        '201':
          description: Created Object Instance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectInstance'
          headers:
            ETag:
              schema:
                type: string
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
    parameters:
      - name: ontologyId
        in: path
        required: true
        schema:
          type: string
          maxLength: 32
      - name: objectTypeId
        in: path
        required: true
        schema:
          type: string
          maxLength: 32
  /v1/ontologies/{ontologyId}/object-types/{objectTypeId}/object-instances/{objectInstanceId}:
    get:
      tags:
        - ObjectInstance
      summary: Get an Object Instance
      operationId: getObjectInstance
      responses:
        '200':
          description: Object Instance
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectInstance'
    patch:
      tags:
        - ObjectInstance
      summary: Update an Object Instance
      description: Merge patch; null removes an override and the primary key is immutable.
      operationId: updateObjectInstance
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateObjectInstanceReq'
      responses:
        '200':
          description: Updated Object Instance
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectInstance'
    delete:
      tags:
        - ObjectInstance
      summary: Delete an Object Instance
      operationId: deleteObjectInstance
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '204':
          description: Deleted Object Instance
    parameters:
      - name: ontologyId
        in: path
        required: true
        schema:
          type: string
          maxLength: 32
      - name: objectTypeId
        in: path
        required: true
        schema:
          type: string
          maxLength: 32
      - name: objectInstanceId
        in: path
        required: true
        schema:
          type: string
          maxLength: 32
...
  /v1/ontologies/{ontologyId}/object-sets/aggregate:
    post:
      tags:
        - Explorer
      summary: Aggregate
      operationId: aggregate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AggregateRequest'
        required: true
      responses:
        '200':
          description: Aggregate response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AggregateResponse'
    parameters:
      - $ref: '#/components/parameters/OntologyId'
...
  /v1/ontologies/{ontologyId}/object-types/{objectTypeId}/object-instances/query:
    post:
      tags:
        - ObjectInstance
      summary: Query Object Instances
      operationId: queryObjectInstances
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ObjectInstanceQueryReq'
      responses:
        '200':
          description: Cursor page of Object Instances
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectInstancePage'
    parameters:
      - $ref: '#/components/parameters/ObjectTypeId'
      - $ref: '#/components/parameters/OntologyId'
  /v1/ontologies/{ontologyId}/object-types/{objectTypeId}/object-instances/aggregate:
    post:
      tags:
        - ObjectInstance
      summary: Aggregate Object Instances
      operationId: aggregateObjectInstances
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ObjectInstanceAggregateReq'
      responses:
        '200':
          description: Aggregate rows
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectInstanceAggregateResult'
    parameters:
      - $ref: '#/components/parameters/ObjectTypeId'
      - $ref: '#/components/parameters/OntologyId'
  /v1/ontologies/{ontologyId}/object-types/{objectTypeId}/object-instances/bulk:
    post:
      tags:
        - ObjectInstance
      summary: Mutate Object Instances in Bulk
      operationId: bulkMutateObjectInstances
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkObjectInstanceReq'
      responses:
        '200':
          description: Per-item mutation results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkObjectInstanceResult'
    parameters:
      - $ref: '#/components/parameters/ObjectTypeId'
      - $ref: '#/components/parameters/OntologyId'
  /v1/ontologies/{ontologyId}/object-types/{objectTypeId}/object-instances/imports:
    post:
      tags:
        - ObjectInstance
      summary: Import a Dataset into an Object Type
      operationId: createObjectInstanceImport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateObjectInstanceImportReq'
      responses:
        '202':
          description: Import accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectInstanceImportJob'
    parameters:
      - $ref: '#/components/parameters/ObjectTypeId'
      - $ref: '#/components/parameters/OntologyId'
  /v1/ontologies/{ontologyId}/object-types/{objectTypeId}/object-instances/imports/{jobId}:
    get:
      tags:
        - ObjectInstance
      summary: Get Object Instance Import
      operationId: getObjectInstanceImport
      responses:
        '200':
          description: Import status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectInstanceImportJob'
    delete:
      tags:
        - ObjectInstance
      summary: Cancel Object Instance Import
      operationId: cancelObjectInstanceImport
      responses:
        '202':
          description: Cancellation requested
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectInstanceImportJob'
    parameters:
      - name: jobId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/ObjectTypeId'
      - $ref: '#/components/parameters/OntologyId'
  /v1/ontologies/{ontologyId}/object-types/{objectTypeId}/object-instances/imports/{jobId}/errors:
    get:
      tags:
        - ObjectInstance
      summary: List Object Instance Import Errors
      operationId: listObjectInstanceImportErrors
      responses:
        '200':
          description: Safe row-level import errors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ObjectInstanceImportError'
    parameters:
      - name: jobId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/ObjectTypeId'
      - $ref: '#/components/parameters/OntologyId'
  /v1/ontologies/{ontologyId}/object-types/{objectTypeId}/object-instances/{objectId}/projection-status:
    get:
      tags:
        - ObjectInstance
      summary: Get Object Instance Projection Status
      operationId: getObjectInstanceProjectionStatus
      responses:
        '200':
          description: Per-target projection versions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectInstanceProjectionStatus'
    parameters:
      - name: objectId
        in: path
        required: true
        schema:
          type: string
          maxLength: 512
      - $ref: '#/components/parameters/ObjectTypeId'
      - $ref: '#/components/parameters/OntologyId'
  /v1/ontologies/{ontologyId}/object-types/{objectTypeId}/object-instances/reconciliations:
    post:
      tags:
        - ObjectInstance
      summary: Reconcile Object Instance Projections
      operationId: reconcileObjectInstances
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReconcileObjectInstancesReq'
      responses:
        '202':
          description: Reconciliation accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectInstanceReconciliationJob'
    parameters:
      - $ref: '#/components/parameters/ObjectTypeId'
      - $ref: '#/components/parameters/OntologyId'
  /v1/ontologies/{ontologyId}/object-types/{objectTypeId}/object-instances/reconciliations/{jobId}:
    get:
      tags:
        - ObjectInstance
      summary: Get Object Instance Reconciliation
      operationId: getObjectInstanceReconciliation
      responses:
        '200':
          description: Reconciliation status and differences
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectInstanceReconciliationResult'
    parameters:
      - name: jobId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/ObjectTypeId'
      - $ref: '#/components/parameters/OntologyId'
...
components:
  schemas:
    CreateOntologyReq:
      type: object
      additionalProperties: false
      required:
        - name
        - id
      properties:
        id:
          type: string
          maxLength: 32
        name:
          type: string
          minLength: 1
          maxLength: 32
        description:
          type: string
          maxLength: 1024
...
    OntologyPage:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            "$ref": "#/components/schemas/Ontology"
        total:
          type: integer
          format: int64
        limit:
          type: integer
          format: int32
        offset:
          type: integer
          format: int32
    Ontology:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - createdAt
        - updatedAt
...
    CreateObjectInstanceReq:
      type: object
      additionalProperties: false
      required:
        - properties
      properties:
        properties:
          type: object
          description: Values keyed by Object Type property names; idField determines the ID and nameField determines the display name.
          additionalProperties: true
    ObjectInstance:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
        title:
          type: string
        objectTypeId:
          type: string
        properties:
          type: object
          additionalProperties: true
          description: Properties keyed by their published business-facing display names.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        type:
          $ref: '#/components/schemas/ResourceId'
        version:
          type: integer
          format: int64
        correlationId:
          type: string
          format: uuid
        projectionStatus:
          $ref: '#/components/schemas/ObjectInstanceProjectionStatus'
      required:
        - id
        - title
        - type
        - properties
        - version
        - createdAt
        - updatedAt
    ObjectInstancePage:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ObjectInstance'
        total:
          type: integer
          format: int64
        limit:
          type: integer
          format: int32
        offset:
          type: integer
          format: int32
        type:
          $ref: '#/components/schemas/ResourceId'
        totalIsLowerBound:
          type: boolean
        nextCursor:
          type:
            - string
            - 'null'
      required:
        - type
        - items
        - total
        - totalIsLowerBound
    UpdateObjectInstanceReq:
      type: object
      additionalProperties: false
      required:
        - properties
      properties:
        properties:
          type: object
          description: Values keyed by Object Type property names; idField cannot be changed.
          additionalProperties: true
...
    ObjectInstanceQueryReq:
      type: object
      additionalProperties: false
      properties:
        properties:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/ObjectInstancePropertyName'
        filters:
          type: array
          maxItems: 50
          items:
            $ref: '#/components/schemas/InstanceFilter'
        filterOperator:
          type: string
          description: Boolean operator used to combine all filters.
          enum:
            - AND
            - OR
          default: AND
        sort:
          type: array
          maxItems: 5
          items:
            $ref: '#/components/schemas/InstanceSort'
        pageSize:
          type: integer
          enum:
            - 25
            - 50
            - 100
          default: 50
        cursor:
          type: string
    ObjectInstanceAggregateReq:
      type: object
      additionalProperties: false
      required:
        - metrics
      properties:
        filters:
          type: array
          maxItems: 50
          items:
            $ref: '#/components/schemas/InstanceFilter'
        filterOperator:
          type: string
          description: Boolean operator used to combine all filters.
          enum:
            - AND
            - OR
          default: AND
        groupBy:
          type: array
          maxItems: 5
          uniqueItems: true
          items:
            $ref: '#/components/schemas/ObjectInstancePropertyName'
        metrics:
          type: array
          minItems: 1
          maxItems: 20
          items:
            $ref: '#/components/schemas/ObjectInstanceAggregateMetric'
    ObjectInstanceAggregateMetric:
      type: object
      additionalProperties: false
      required:
        - operation
      properties:
        operation:
          type: string
          enum:
            - AVG
            - COUNT
            - DISTINCT_COUNT
            - MAX
            - MIN
            - SUM
        property:
          $ref: '#/components/schemas/ObjectInstancePropertyName'
        alias:
          type: string
          maxLength: 160
    ObjectInstanceAggregateResult:
      type: object
      additionalProperties: false
      required:
        - rows
      properties:
        rows:
          type: array
          items:
            type: object
            additionalProperties: true
    BulkObjectInstanceReq:
      type: object
      additionalProperties: false
      required:
        - operation
        - items
      properties:
        operation:
          type: string
          enum:
            - CREATE
            - DELETE
            - UPDATE
            - UPSERT
        atomic:
          type: boolean
          default: false
        items:
          type: array
          minItems: 1
          maxItems: 500
          items:
            $ref: '#/components/schemas/BulkObjectInstanceItem'
    BulkObjectInstanceItem:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          maxLength: 512
        version:
          type: integer
          format: int64
          minimum: 1
        properties:
          type: object
          additionalProperties: true
    BulkObjectInstanceResult:
      type: object
      additionalProperties: false
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/BulkObjectInstanceItemResult'
    BulkObjectInstanceItemResult:
      type: object
      additionalProperties: false
      required:
        - index
        - status
      properties:
        index:
          type: integer
          format: int32
        id:
          type: string
        status:
          type: string
          enum:
            - CREATED
            - DELETED
            - FAILED
            - UNCHANGED
            - UPDATED
        version:
          type: integer
          format: int64
        errorCode:
          type: string
        safeError:
          type: string
    CreateObjectInstanceImportReq:
      type: object
      additionalProperties: false
      required:
        - datasetId
        - identityField
        - titleField
        - fieldMappings
      properties:
        datasetId:
          $ref: '#/components/schemas/ResourceId'
        identityField:
          type: string
        titleField:
          type: string
        fieldMappings:
          type: object
          description: Map from Dataset field names to published Object Type property display names.
          additionalProperties:
            type: string
        mode:
          type: string
          enum:
            - REPLACE
            - UPSERT
          default: UPSERT
    ObjectInstanceImportJob:
      type: object
      additionalProperties: false
      required:
        - id
        - status
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - CANCELLED
            - COMPLETED
            - FAILED
            - MERGING
            - PARTIAL
            - QUEUED
            - VALIDATING
        inserted:
          type: integer
          format: int64
        updated:
          type: integer
          format: int64
        deleted:
          type: integer
          format: int64
        unchanged:
          type: integer
          format: int64
        failed:
          type: integer
          format: int64
        safeError:
          type: string
        createdAt:
          type: string
          format: date-time
        completedAt:
          type:
            - string
            - 'null'
          format: date-time
    ObjectInstanceImportError:
      type: object
      additionalProperties: false
      required:
        - rowNumber
        - errorCode
        - safeMessage
      properties:
        rowNumber:
          type: integer
          format: int64
        objectId:
          type: string
        fieldId:
          type: string
        errorCode:
          type: string
        safeMessage:
          type: string
    ObjectInstanceProjectionTargetStatus:
      type: object
      additionalProperties: false
      required:
        - target
        - version
        - status
      properties:
        target:
          type: string
          enum:
            - HUGEGRAPH
            - OPENSEARCH
        version:
          type: integer
          format: int64
        status:
          type: string
        lastError:
          type: string
        updatedAt:
          type: string
          format: date-time
    ObjectInstanceProjectionStatus:
      type: object
      additionalProperties: false
      required:
        - objectId
        - authoritativeVersion
        - targets
      properties:
        objectId:
          type: string
        authoritativeVersion:
          type: integer
          format: int64
        correlationId:
          type: string
          format: uuid
        targets:
          type: array
          items:
            $ref: '#/components/schemas/ObjectInstanceProjectionTargetStatus'
    ReconcileObjectInstancesReq:
      type: object
      additionalProperties: false
      properties:
        repair:
          type: boolean
          default: true
    ObjectInstanceReconciliationJob:
      type: object
      additionalProperties: false
      required:
        - id
        - status
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
        missing:
          type: integer
          format: int64
        stale:
          type: integer
          format: int64
        extra:
          type: integer
          format: int64
        repaired:
          type: integer
          format: int64
...
    ObjectInstanceReconciliationResult:
      type: object
      additionalProperties: false
      required:
        - job
        - differences
      properties:
        job:
          $ref: '#/components/schemas/ObjectInstanceReconciliationJob'
        differences:
          type: array
          items:
            $ref: '#/components/schemas/ObjectInstanceReconciliationDifference'
    ObjectInstancePropertyName:
      type: string
      description: Published business-facing display name of an Object Type property.
      minLength: 1
      maxLength: 240
    AggregateRequest:
      type: object
      properties:
        measurePropertyId:
          type: string
          format: uuid
        divisorPropertyId:
          type: string
          format: uuid
        query:
          $ref: '#/components/schemas/ObjectSetRequest'
        dimensionPropertyIds:
          type: array
          items:
            type: string
            format: uuid
        aggregation:
          type: string
    AggregateResponse:
      type: object
      properties:
        objectTypeId:
          type: string
          format: uuid
        queryFingerprint:
          type: string
        buckets:
          type: array
          items:
            $ref: '#/components/schemas/AggregationBucket'
        calculatedAt:
          type: string
          format: date-time
...
    ObjectSetRequest:
      type: object
      properties:
        objectTypeId:
          type: string
          format: uuid
        where:
          type: object
          additionalProperties: {}
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortClause'
        pageSize:
          type: integer
          format: int32
        cursor:
          type: string
        columns:
          type: array
          items:
            type: string
            format: uuid
...
    ResourceId:
      type: string
      description: Caller-defined stable ID used in API paths and references
      example: employee
      pattern: ^[A-Za-z][A-Za-z0-9_-]{0,159}$
...
    AggregationBucket:
      type: object
      properties:
        value: {}
        count:
          type: integer
          format: int64
        metric:
          type: number
          format: double
...
    InstanceFilter:
      type: object
      additionalProperties: false
      required:
        - property
        - operator
      properties:
        property:
          $ref: '#/components/schemas/ObjectInstancePropertyName'
        operator:
          type: string
          enum:
            - eq
            - ne
            - gt
            - gte
            - lt
            - lte
            - between
            - in
            - not_in
            - exact
            - starts_with
            - is_empty
            - is_not_empty
        value: {}
...
    InstanceSort:
      type: object
      additionalProperties: false
      required:
        - property
        - direction
      properties:
        property:
          $ref: '#/components/schemas/ObjectInstancePropertyName'
        direction:
          type: string
          enum:
            - asc
            - desc
...
    SortClause:
      type: object
      properties:
        propertyId:
          type: string
          format: uuid
        direction:
          type: string
...
  parameters:
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
...
    OntologyId:
      name: ontologyId
      in: path
      description: Caller-defined stable Ontology ID
      required: true
      schema:
        $ref: '#/components/schemas/ResourceId'
    IfMatch:
      name: If-Match
      in: header
      description: ETag returned by the latest read of the resource
      required: true
      schema:
        type: string
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: Stable key for safely replaying a create or bulk request
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 240
    ObjectTypeId:
      name: objectTypeId
      in: path
      description: Caller-defined stable Object Type ID
      required: true
      schema:
        $ref: '#/components/schemas/ResourceId'
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        enum:
          - 25
          - 50
          - 100
        default: 50